// client/src/store/mutations/types/recipeTypes.ts
export enum RecipeMutationType {
  SET_RECIPES = "SET_RECIPES",
  ADD_RECIPE = "ADD_RECIPE",
  CLONE_RECIPE = "CLONE_RECIPE",
  CLEAR_CLONE_RECIPE = "CLEAR_CLONE_RECIPE",
  UPDATE_RECIPE = "UPDATE_RECIPE",
  DELETE_RECIPE = "DELETE_RECIPE",
  SET_SELECTED = "SET_SELECTED",
  CLEAR_SELECTED = "CLEAR_SELECTED",
}